requirements.txt
aiohttp
aiosignal
argcomplete
asgiref
async-timeout
attrs
boto
CacheControl
cachetools
certifi
cffi
charset-normalizer
ci-info
click
cloudinary
configobj
configparser
crcmod
cryptography
defusedxml
dj-database-url
Django
django-admin
django-cloudinary-storage
django-cors-headers
django-excel-response2
django-six
djangorestframework
etelemetry
excel-base
fasteners
filelock
firebase-admin
fonttools
fpdf2
frozenlist
gcs-oauth2-boto-plugin
google-api-core
google-api-python-client
google-apitools
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-reauth
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
gsutil
gunicorn
httplib2
idna
isodate
isoweek
looseversion
lxml
monotonic
msgpack
multidict
networkx
nibabel
nipype
numpy
oauth2client
opencv-python
packaging
pandas
pathlib
pillow
proto-plus
protobuf
prov
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydot
PyJWT
PyMuPDF
PyMuPDFb
pyOpenSSL
pyparsing
PyPDF2
python-dateutil
python-dotenv
pytz
pyu2f
pyxnat
rdflib
requests
retry-decorator
rsa
scipy
screen
simplejson
six
sqlparse
TimeConvert
traits
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
xlwt
yarl
